test: tests to match updated regex
